webcave: Move compose file to repo root
Move compose file for webcave server to repo root since it is needed by the  infra role with which we will deploy it.
Updated the compose file to be compatible with environment variable forwarding from the ansible role. Namely, we need
PORT, IS_BEHIND_PROXY and CLIENT_ORIGIN_URL to be read from environment of the process starting the containers. Also,
in the case these values are not present or empty, just use the default ones.
Update Readme file to reflect changes.

Referenced issue: https://github.com/status-im/infra-misc/issues/308

Signed-off-by: markoburcul <marko@status.im>